+++ plugins/11_TheWarWithin/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v121
+
+* Removed [I TAKE Candle!](https://www.wowhead.com/achievement=40446/i-take-candle) from Kobold delves as achievement is now a Feat of Strength and can no longer be attained.
+
+# v120
+
+* Add Nightfall elite locations for [Fighter of the Nightman](https://www.wowhead.com/achievement=41999/fighter-of-the-nightman)
+* Fix note for [Back to the Wall](https://www.wowhead.com/achievement=40620/back-to-the-wall)
+* Localization updates: deDE, esES, frFR, ruRU, zhCN
+
 # v119
 
 * Fix lua error when hovering scrap heaps in undermine [#492]